+++ lib/src/model/arrow_horizontal_direction.dart
@@ -1,0 +1,11 @@
+/// 控制横向箭头方向的枚举类
+/// Enum class to control horizontal arrow direction
+enum ArrowHorizontalDirection {
+  /// 向左
+  /// Arrow pointing left
+  left,
+
+  /// 向右
+  /// Arrow pointing right
+  right,
+}